Implement PmodOLED having an address in AMBA 3
We've somehow managed to forget about the fact that each device has an
address as well. So we've added a new structure to hold it and modified
the infrastructure to pass it everywhere.

+++ drv/pmodoled.h
@@ -12,7 +12,7 @@ struct pmodoled_config {
 void pmodoled_config_init(struct pmodoled_config *config);
 
 struct pmodoled_drv {
-        int (*write_word)(void *, uint32_t addr, uint32_t value);
+        int (*write_word)(void *, uint16_t addr, uint32_t value);
         void *data;
         struct pmodoled_config config;
 };
